refactor: unify data key inference and complete context resolution
